requirements.txt
psycopg2-binary==2.9.3
SQLAlchemy==1.4.40
sqlparse==0.4.2